fix an integer overflow issue with SD cards >4GB (cast getAvailableBlocks to long)
git-svn-id: http://five.googlecode.com/svn/trunk/five-client@836 0f5e3bc8-7f4c-0410-bf3f-39b466953847

+++ src/org/devtcg/five/service/CacheService.java
@@ -107,11 +107,9 @@ public class CacheService extends Service
 OUTER:
 			while (true)
 			{
-				/* XXX: Currently undocumented interface, but we need the
-				 * punch through to stat(). */
 				StatFs fs = new StatFs(sdcard.getAbsolutePath());
-
-				long freeBytes = fs.getAvailableBlocks() * fs.getBlockSize();
+				
+				long freeBytes = (long)fs.getAvailableBlocks() * fs.getBlockSize();
 				long necessary = mPolicy.leaveFree - (freeBytes + size);
 
 				if (necessary <= 0)
@@ -189,7 +187,7 @@ OUTER:
 
 			return path;
 		}
-		
+
 		public ParcelFileDescriptor requestStorage(long sourceId,
 		  long contentId)
 		  throws RemoteException
